--- nft/test_bootstrapper.go
@@ -1,11 +1,0 @@
-// +build integration unit
-
-package nft
-
-func (b *Bootstrapper) TestBootstrap(context map[string]interface{}) error {
-	return b.Bootstrap(context)
-}
-
-func (b *Bootstrapper) TestTearDown() error {
-	return nil
-}
