+++ testworld/nft_test.go
@@ -7,7 +7,8 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/centrifuge/go-centrifuge/documents"
+	"github.com/centrifuge/go-centrifuge/httpapi/coreapi"
+	"github.com/centrifuge/go-centrifuge/identity"
 	"github.com/centrifuge/go-centrifuge/nft"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/gavv/httpexpect"
@@ -25,7 +26,10 @@ func defaultNFTMint(t *testing.T, documentType string) nft.TokenID {
 	assetAddress := common.HexToAddress(alice.host.dappAddresses["assetManager"])
 
 	// Alice shares document with Bob
-	res := createDocument(alice.httpExpect, alice.id.String(), documentType, http.StatusAccepted, genericCoreAPICreate([]string{bob.id.String()}))
+	docPayload := genericCoreAPICreate([]string{bob.id.String()})
+	attrs, pfs := getAttributeMapRequest(t, alice.id)
+	docPayload["attributes"] = attrs
+	res := createDocument(alice.httpExpect, alice.id.String(), documentType, http.StatusAccepted, docPayload)
 	txID := getTransactionID(t, res)
 	status, message := getTransactionStatusAndMessage(alice.httpExpect, alice.id.String(), txID)
 	if status != "success" {
@@ -37,8 +41,8 @@ func defaultNFTMint(t *testing.T, documentType string) nft.TokenID {
 		t.Error("docIdentifier empty")
 	}
 
-	getGenericDocumentAndCheck(t, alice.httpExpect, alice.id.String(), docIdentifier, nil, createAttributes())
-	getGenericDocumentAndCheck(t, bob.httpExpect, bob.id.String(), docIdentifier, nil, createAttributes())
+	getGenericDocumentAndCheck(t, alice.httpExpect, alice.id.String(), docIdentifier, nil, attrs)
+	getGenericDocumentAndCheck(t, bob.httpExpect, bob.id.String(), docIdentifier, nil, attrs)
 
 	var response *httpexpect.Object
 	var err error
@@ -46,11 +50,13 @@ func defaultNFTMint(t *testing.T, documentType string) nft.TokenID {
 	depositAddress := alice.id.String()
 
 	// mint an NFT
+	acr, err := alice.host.configService.GetAccount(alice.id.ToAddress().Bytes())
+	pfs = append(pfs, nft.GetSignatureProofField(t, acr))
 	payload := map[string]interface{}{
 		"document_id":           docIdentifier,
 		"registry_address":      registry.String(),
 		"deposit_address":       depositAddress, // Centrifuge address
-		"proof_fields":          []string{documents.CDTreePrefix + ".next_version"},
+		"proof_fields":          pfs,
 		"asset_manager_address": assetAddress,
 	}
 	response, err = alice.host.mintNFT(alice.httpExpect, alice.id.String(), http.StatusAccepted, payload)
@@ -62,7 +68,7 @@ func defaultNFTMint(t *testing.T, documentType string) nft.TokenID {
 		t.Error(message)
 	}
 
-	docVal := getGenericDocumentAndCheck(t, alice.httpExpect, alice.id.String(), docIdentifier, nil, createAttributes())
+	docVal := getGenericDocumentAndCheck(t, alice.httpExpect, alice.id.String(), docIdentifier, nil, attrs)
 	assert.True(t, len(docVal.Path("$.header.nfts[0].token_id").String().Raw()) > 0, "successful tokenId should have length 77")
 	assert.True(t, len(docVal.Path("$.header.nfts[0].token_index").String().Raw()) > 0, "successful tokenIndex should have a value")
 
@@ -75,42 +81,6 @@ func defaultNFTMint(t *testing.T, documentType string) nft.TokenID {
 	assert.Equal(t, strings.ToLower(depositAddress), strings.ToLower(owner.Hex()))
 	assert.Equal(t, strings.ToLower(respOwner), strings.ToLower(owner.Hex()))
 	return tokenID
-}
-
-func TestInvoiceUnpaidMint_errors(t *testing.T) {
-	t.SkipNow() //TODO enable as soon as we have adapted NFT invoice unpaid
-	t.Parallel()
-	alice := doctorFord.getHostTestSuite(t, "Alice")
-	tests := []struct {
-		errorMsg   string
-		httpStatus int
-		payload    map[string]interface{}
-	}{
-		{
-
-			"RegistryAddress is not a valid Ethereum address",
-			http.StatusBadRequest,
-			map[string]interface{}{
-				"registry_address": "0x123",
-			},
-		},
-		{
-			"cannot unmarshal hex string without 0x",
-			http.StatusBadRequest,
-			map[string]interface{}{
-				"registry_address": "0xf72855759a39fb75fc7341139f5d7a3974d4da08", //dummy address
-				"deposit_address":  "abc",
-			},
-		},
-	}
-	for _, test := range tests {
-		t.Run(test.errorMsg, func(t *testing.T) {
-			t.Parallel()
-			response, err := alice.host.mintNFT(alice.httpExpect, alice.id.String(), test.httpStatus, test.payload)
-			assert.Nil(t, err, "it should be possible to call the API endpoint")
-			response.Value("message").String().Contains(test.errorMsg)
-		})
-	}
 }
 
 func TestTransferNFT_successful(t *testing.T) {
@@ -148,4 +118,18 @@ func TestTransferNFT_successful(t *testing.T) {
 	resp, err = alice.host.ownerOfNFT(alice.httpExpect, alice.id.String(), http.StatusOK, ownerOfPayload)
 	assert.NoError(t, err)
 	resp.Path("$.owner").String().Equal(strings.ToLower(bob.id.String()))
+}
+
+func getAttributeMapRequest(t *testing.T, did identity.DID) (coreapi.AttributeMapRequest, []string) {
+	attrs, pfs := nft.GetAttributes(t, did)
+	amr := make(coreapi.AttributeMapRequest)
+	for _, attr := range attrs {
+		v, err := attr.Value.String()
+		assert.NoError(t, err)
+		amr[attr.KeyLabel] = coreapi.AttributeRequest{
+			Type:  attr.Value.Type.String(),
+			Value: v,
+		}
+	}
+	return amr, pfs
 }
